Fixed frame delay plus resize bug
Former-commit-id: 32ce81eb0e5d90d0b4f2a47e967dfed4c7ec3556
Former-commit-id: d243591041847d982488e17fecdc6e2c76d866b0

+++ src/ImageProcessor/Imaging/Formats/GifInfo.cs
@@ -53,6 +53,6 @@ namespace ImageProcessor.Imaging.Formats
         /// <summary>
         /// Gets or sets the animation length in milliseconds.
         /// </summary>
-        public int AnimationLength { get; set; }
+        public double AnimationLength { get; set; }
     }
 }
